migrate hashauth to class + jsdoc
Since this was previously exported as an *instance* rather than a function returning an instance, I've had to change the signature of `hashauth.js`. This isn't too bad - we lose the propety `hashauth.initialized` (which was never used), and now instantiate the class rather than calling `.init(client, $)`.
Also, a few tests used `hashauth` for a reason I can't figure out. These tests pass without the `hashauth` part both on this branch and on `master`