Merge upstream changes, use them in `frontend` dir (#3)
* Added notes about the Google Assistant integration being broken

* Fix missing mmol unit conversion for lastEnacted.bg

* allow tests to pass

Switching the way units is fetched to this way makes it similar to the other
code already in place that does something similar.  This allows the tests to
pass with the existing fixtures.
Considering augmenting  additional tests that explicitly tests whether the
relevant branch of code is taken.

* permit any LTS version of node

* continuously test on lts also

This adds the latest long term support version of node to the automated testing
matrix.

* setup-node action syntax is lts/*

* use upstream changes in `frotnend` dir

---------

Co-authored-by: Caleb <inventor96@users.noreply.github.com>
Co-authored-by: inventor96 <inventor96@gmail.com>
Co-authored-by: Ben West <bewest@gmail.com>
Co-authored-by: dsnallfot <72826201+dsnallfot@users.noreply.github.com>