Migrate loop plugin
Rewrote basically the entirety of `analyzeData` to be declarative - using `filter` and `map` and `reduce` rather than mutable properties on an object.
Also rewrote `updateVisualisation` to be declarative (except that we push to `events` a lot). Moved the local functions to protected instance methods, with each making a new array, rather than taking an existing one as an arg and concatenating onto that.

Rewrote the virtAsst handlers and the `status${"Level" | "Class"}` methods to be more declarative (just return, don't mutate a variable then return that at the end)