migrate SensorAgePlugin to class + jsdoc
Some weirdness with the types: `findLatestTimeChange` returns an object with 2 keys containing data, and the third key, `min` telling us which of the two data fields to look at.
The data field we *should* look at has `.notificastion?: Notify` and `.level: Level`. Unfortunately, despite the narrowest types possible, `sensorInfos[sensorInfos.min]` does not provide the correct types, so we have to do a funky ternary to appease typescript