fix ar2 translationkey/string type error
In part due to the fact that `High`/`Low` _are_ translationkeys, but `high`/`low` aren't. The `.translate` function is case insensitive as loing as you don't pass it any other args (string interpolation), so this doesn't matter most of the time. The performance impacts of doing a case insensitive `TranslationKeys` type is not worth it (though technically possible)