migrate `xdripjs` plugin to class + jsdoc
Mostly farily standard - the `getStateString`/`findLatestState` function was renamed to `analyzeData` and made more declarative, `updateVisualisation` left mostly untouched.
The `virtAsst` stuff was a little funky - especially with types. I changes the signature so the `virtAsstGenericCGMHandler` is now a factory which returns a `VirtAsstIntentHandlerFn`. Its arguments are a string to append to `virtAsstTitleCGM` *and* `virtAsstCGM`, which is typed correctly using `RemovePrefix` on all the `TranslationKeys` such that both `virtAsstCGM` *and* `virtAsstTitleCGM` when prefixed on the argument, is a valid `TranslationKey`.
The second argument is the name of a field on the plugin'ls `properties` which is stringifiable in place. I did this by extracting all the field names which map to strings or numbers. I then had to union this with the names of fields which mapped to `Level`s, since `Level` is more specific than number, so `string | number extends Level` is false. But since `number | Level = number`, we can't do this union within the `KeysOfType` call.
If `Level` were a branded type, this would not be an issue